refactor: コントローラ構造整理（Community新設 + API整理） (#394)
- CommunityController 新設（PlansController#indexから分離）
- InfoWindowsController を API から トップレベルに移動
- API コントローラを plans/ 配下に整理
- adopt アクションを AdoptionsController に分離
