+++ app/controllers/plans_controller.rb
@@ -1,29 +1,5 @@
 class PlansController < ApplicationController
-  before_action :authenticate_user!, except: %i[index show]
-
-  def index
-    set_filter_variables
-
-    # お気に入りフィルターが有効な場合、current_user を渡す（ログイン時のみ）
-    liked_by_user = @favorites_only && current_user ? current_user : nil
-
-    # 検索タイプに応じてプランまたはスポットを取得
-    if @search_type == "spot"
-      @community_spots = Spot.for_community(
-        keyword: params[:q],
-        cities: params[:cities],
-        genre_ids: params[:genre_ids],
-        liked_by_user: liked_by_user
-      ).page(params[:page]).per(10)
-    else
-      @plans = Plan.for_community(
-        keyword: params[:q],
-        cities: params[:cities],
-        genre_ids: params[:genre_ids],
-        liked_by_user: liked_by_user
-      ).page(params[:page]).per(10)
-    end
-  end
+  before_action :authenticate_user!, except: %i[show]
 
   def show
     @plan = Plan.publicly_visible
@@ -47,30 +23,6 @@ class PlansController < ApplicationController
 
   def edit
     @plan = current_user.plans.includes(:start_point, :goal_point, plan_spots: :spot).find(params[:id])
-
-    set_filter_variables
-
-    # お気に入りフィルターが有効な場合、current_user を渡す
-    liked_by_user = @favorites_only ? current_user : nil
-
-    # 検索タイプに応じてプランまたはスポットを取得
-    if @search_type == "spot"
-      @community_spots = Spot.for_community(
-        keyword: params[:q],
-        cities: params[:cities],
-        genre_ids: params[:genre_ids],
-        liked_by_user: liked_by_user
-      ).page(params[:page]).per(10)
-    else
-      @community_plans = Plan.for_community(
-        keyword: params[:q],
-        cities: params[:cities],
-        genre_ids: params[:genre_ids],
-        liked_by_user: liked_by_user
-      ).where.not(id: @plan.id)
-        .page(params[:page])
-        .per(6)
-    end
   end
 
   def update
@@ -92,22 +44,12 @@ class PlansController < ApplicationController
   def destroy
     @plan = current_user.plans.find(params[:id])
     @plan.destroy!
-    redirect_back fallback_location: plans_path, notice: "プランを削除しました"
+    redirect_back fallback_location: community_path, notice: "プランを削除しました"
   end
 
   private
 
   def plan_params
     params.require(:plan).permit(:title)
-  end
-
-  def set_filter_variables
-    @search_type = params[:search_type]
-    @search_query = params[:q]
-    @selected_cities = Array(params[:cities]).reject(&:blank?)
-    @selected_genre_ids = Array(params[:genre_ids]).map(&:to_i).reject(&:zero?)
-    @favorites_only = params[:favorites_only] == "1"
-    @genres_by_category = Genre.grouped_by_category
-    @cities_by_prefecture = Spot.cities_by_prefecture
   end
 end
