+++ app/views/community/_results.html.erb
@@ -1,30 +1,35 @@
-<%# みんなのプラン一覧（plans#index） %>
+<%# コミュニティ検索結果（standalone / Turbo Frame 共用） %>
+<% turbo_frame = local_assigns.fetch(:turbo_frame, false) %>
 
-<div class="list-page list-page--two-column">
-  <h1 class="list-page__title">みんなの旅</h1>
+<%= turbo_frame_tag "community_content", target: (turbo_frame ? "_self" : "_top") do %>
+  <% if turbo_frame %>
+    <%# Turbo Frame 時は検索フォームも含める %>
+    <div class="list-page">
+      <%= render "shared/search_form",
+                 form_url: community_path,
+                 turbo_frame: "community_content",
+                 search_type: @search_type,
+                 selected_cities: @selected_cities,
+                 cities_by_prefecture: @cities_by_prefecture,
+                 genres_by_category: @genres_by_category,
+                 selected_genre_ids: @selected_genre_ids,
+                 search_query: @search_query,
+                 show_favorites_toggle: true,
+                 favorites_only: @favorites_only,
+                 button_class: "search-button--blue" %>
 
-  <div class="list-page__content">
-  <%# 左サイドバー: 検索フォーム %>
-  <div class="list-page__sidebar">
-    <%= render "shared/search_form",
-               form_url: plans_path,
-               search_type: @search_type,
-               selected_cities: @selected_cities,
-               cities_by_prefecture: @cities_by_prefecture,
-               genres_by_category: @genres_by_category,
-               selected_genre_ids: @selected_genre_ids,
-               search_query: @search_query %>
-  </div>
+      <div class="results-divider">
+        <span class="results-label">検索結果</span>
+      </div>
+  <% end %>
 
-  <%# 右メイン: カード一覧 + ページネーション %>
-  <div class="list-page__main">
   <% if @search_type == "spot" %>
     <%# スポット検索結果 %>
     <% if @community_spots.present? %>
       <% favorite_spots_map = FavoriteSpot.index_by_spot_id(user: current_user, spot_ids: @community_spots.map(&:id)) %>
       <div class="list-page__cards">
         <% @community_spots.each do |spot| %>
-          <%= render "plans/spot_card", spot: spot, favorite_spot: favorite_spots_map[spot.id] %>
+          <%= render "plans/spot_card", spot: spot, favorite_spot: favorite_spots_map[spot.id], preview_mode: turbo_frame %>
         <% end %>
       </div>
 
@@ -45,22 +50,23 @@
     <% end %>
   <% else %>
     <%# プラン検索結果 %>
-    <% if @plans.present? %>
-      <% favorite_plans_map = FavoritePlan.index_by_plan_id(user: current_user, plan_ids: @plans.map(&:id)) %>
+    <% if @community_plans.present? %>
+      <% favorite_plans_map = FavoritePlan.index_by_plan_id(user: current_user, plan_ids: @community_plans.map(&:id)) %>
       <div class="list-page__cards">
-        <% @plans.each do |plan| %>
+        <% @community_plans.each do |plan| %>
           <%
-            # plan_spots を position 順で取得し、_plan_card 用のハッシュ形式に変換
             spots_data = plan.plan_spots.sort_by(&:position).map do |ps|
               spot = ps.spot
               {
                 name: spot.name,
                 address: spot.address,
+                lat: spot.lat,
+                lng: spot.lng,
+                place_id: spot.place_id,
                 prefecture: spot.prefecture,
                 city: spot.city
               }
             end
-            # 全スポットからユニークなジャンル名を収集
             genre_names = plan.plan_spots.flat_map { |ps| ps.spot.genres.map(&:name) }.uniq
           %>
           <%= render "plans/plan_card",
@@ -70,14 +76,15 @@
                      total_duration: format_move_time(plan.spots_only_move_time),
                      total_distance: format_distance(plan.spots_only_distance),
                      plan: plan,
-                     favorite_plan: favorite_plans_map[plan.id] %>
+                     favorite_plan: favorite_plans_map[plan.id],
+                     preview_mode: turbo_frame %>
         <% end %>
       </div>
 
       <%# ページネーション %>
-      <% if @plans.total_pages > 1 %>
+      <% if @community_plans.total_pages > 1 %>
         <div class="list-page__pagination">
-          <%= paginate @plans %>
+          <%= paginate @community_plans %>
         </div>
       <% end %>
     <% else %>
@@ -90,6 +97,8 @@
       </div>
     <% end %>
   <% end %>
-  </div>
-  </div>
-</div>
+
+  <% if turbo_frame %>
+    </div>
+  <% end %>
+<% end %>
