+++ app/views/plans/_community_tab.html.erb
@@ -1,100 +1,3 @@
-<%# みんなのプランタブ: プラン/スポット検索切り替え %>
+<%# みんなのプランタブ: Turbo Frame で CommunityController から読み込み %>
 
-<%= turbo_frame_tag "community_plans" do %>
-  <div class="list-page">
-    <%# 検索フォーム %>
-    <%= render "shared/search_form",
-               form_url: edit_plan_path(@plan),
-               turbo_frame: "community_plans",
-               search_type: @search_type,
-               selected_cities: @selected_cities,
-               cities_by_prefecture: @cities_by_prefecture,
-               genres_by_category: @genres_by_category,
-               selected_genre_ids: @selected_genre_ids,
-               search_query: @search_query,
-               show_favorites_toggle: true,
-               favorites_only: @favorites_only,
-               button_class: "search-button--blue" %>
-
-    <div class="results-divider">
-      <span class="results-label">検索結果</span>
-    </div>
-
-    <% if @search_type == "spot" %>
-      <%# スポット検索結果 %>
-      <% if @community_spots.present? %>
-        <% favorite_spots_map = FavoriteSpot.index_by_spot_id(user: current_user, spot_ids: @community_spots.map(&:id)) %>
-        <div class="list-page__cards">
-          <% @community_spots.each do |spot| %>
-            <%= render "plans/spot_card", spot: spot, favorite_spot: favorite_spots_map[spot.id], preview_mode: true %>
-          <% end %>
-        </div>
-
-        <%# ページネーション %>
-        <% if @community_spots.total_pages > 1 %>
-          <div class="list-page__pagination">
-            <%= paginate @community_spots %>
-          </div>
-        <% end %>
-      <% else %>
-        <div class="list-page__empty">
-          <% if @search_query.present? || @selected_cities.present? || @selected_genre_ids.present? %>
-            <p>条件に一致するスポットが見つかりませんでした</p>
-          <% else %>
-            <p>まだスポットがありません</p>
-          <% end %>
-        </div>
-      <% end %>
-    <% else %>
-      <%# プラン検索結果 %>
-      <% if @community_plans.present? %>
-        <% favorite_plans_map = FavoritePlan.index_by_plan_id(user: current_user, plan_ids: @community_plans.map(&:id)) %>
-        <div class="list-page__cards">
-          <% @community_plans.each do |plan| %>
-            <%
-              # plan_spots を position 順で取得し、_plan_card 用のハッシュ形式に変換
-              spots_data = plan.plan_spots.sort_by(&:position).map do |ps|
-                spot = ps.spot
-                {
-                  name: spot.name,
-                  address: spot.address,
-                  lat: spot.lat,
-                  lng: spot.lng,
-                  place_id: spot.place_id,
-                  prefecture: spot.prefecture,
-                  city: spot.city
-                }
-              end
-              # 全スポットからユニークなジャンル名を収集
-              genre_names = plan.plan_spots.flat_map { |ps| ps.spot.genres.map(&:name) }.uniq
-            %>
-            <%= render "plans/plan_card",
-                       plan_title: plan.title.presence || "無題のプラン",
-                       spots: spots_data,
-                       genres: genre_names,
-                       total_duration: format_move_time(plan.spots_only_move_time),
-                       total_distance: format_distance(plan.spots_only_distance),
-                       plan: plan,
-                       favorite_plan: favorite_plans_map[plan.id],
-                       preview_mode: true %>
-          <% end %>
-        </div>
-
-        <%# ページネーション %>
-        <% if @community_plans.total_pages > 1 %>
-          <div class="list-page__pagination">
-            <%= paginate @community_plans %>
-          </div>
-        <% end %>
-      <% else %>
-        <div class="list-page__empty">
-          <% if @search_query.present? || @selected_cities.present? || @selected_genre_ids.present? %>
-            <p>条件に一致するプランが見つかりませんでした</p>
-          <% else %>
-            <p>まだ公開プランがありません</p>
-          <% end %>
-        </div>
-      <% end %>
-    <% end %>
-  </div>
-<% end %>
+<%= turbo_frame_tag "community_content", src: community_path %>
