+++ spec/requests/infowindows_spec.rb
@@ -2,7 +2,7 @@
 
 require "rails_helper"
 
-RSpec.describe "Api::Infowindows", type: :request do
+RSpec.describe "Infowindows", type: :request do
   let(:user) { create(:user) }
   let(:plan) { create(:plan, user: user) }
   let(:spot) { create(:spot) }
@@ -12,19 +12,19 @@ RSpec.describe "Api::Infowindows", type: :request do
     stub_google_places_api
   end
 
-  describe "GET /api/infowindow" do
+  describe "GET /infowindow" do
     context "ログイン済みの場合" do
       before { sign_in user }
 
       it "spot_idでInfoWindowを表示する" do
-        get api_infowindow_path, params: { spot_id: spot.id, plan_id: plan.id }
+        get infowindow_path, params: { spot_id: spot.id, plan_id: plan.id }
 
         expect(response).to have_http_status(:ok)
         expect(response.body).to include(spot.name)
       end
 
       it "place_idでInfoWindowを表示する" do
-        get api_infowindow_path, params: { place_id: "ChIJtest123", lat: 35.6580, lng: 139.7016 }
+        get infowindow_path, params: { place_id: "ChIJtest123", lat: 35.6580, lng: 139.7016 }
 
         expect(response).to have_http_status(:ok)
       end
@@ -32,7 +32,7 @@ RSpec.describe "Api::Infowindows", type: :request do
       it "edit_mode=start_pointで出発地点用UIを返す" do
         create(:start_point, plan: plan, address: "東京都渋谷区")
 
-        get api_infowindow_path, params: { edit_mode: "start_point", plan_id: plan.id, name: "自宅" }
+        get infowindow_path, params: { edit_mode: "start_point", plan_id: plan.id, name: "自宅" }
 
         expect(response).to have_http_status(:ok)
         expect(response.body).to include("自宅")
@@ -41,7 +41,7 @@ RSpec.describe "Api::Infowindows", type: :request do
       it "edit_mode=goal_pointで帰宅地点用UIを返す" do
         create(:goal_point, plan: plan, address: "東京都新宿区")
 
-        get api_infowindow_path, params: { edit_mode: "goal_point", plan_id: plan.id, name: "帰宅" }
+        get infowindow_path, params: { edit_mode: "goal_point", plan_id: plan.id, name: "帰宅" }
 
         expect(response).to have_http_status(:ok)
         expect(response.body).to include("帰宅")
@@ -50,7 +50,7 @@ RSpec.describe "Api::Infowindows", type: :request do
       it "プラン内のスポットの場合plan_spot_idを含む" do
         plan_spot = create(:plan_spot, plan: plan, spot: spot)
 
-        get api_infowindow_path, params: { spot_id: spot.id, plan_id: plan.id }
+        get infowindow_path, params: { spot_id: spot.id, plan_id: plan.id }
 
         expect(response).to have_http_status(:ok)
         # プランに含まれているスポットの場合、削除ボタンが表示される
@@ -59,26 +59,26 @@ RSpec.describe "Api::Infowindows", type: :request do
 
     context "未ログインの場合" do
       it "ゲスト用UIを返す" do
-        get api_infowindow_path, params: { spot_id: spot.id }
+        get infowindow_path, params: { spot_id: spot.id }
 
         expect(response).to have_http_status(:ok)
         # ゲスト用UIはログインを促すメッセージを含む
       end
 
       it "認証不要でアクセス可能" do
-        get api_infowindow_path, params: { place_id: "ChIJtest123" }
+        get infowindow_path, params: { place_id: "ChIJtest123" }
 
         expect(response).to have_http_status(:ok)
       end
     end
   end
 
-  describe "POST /api/infowindow" do
+  describe "POST /infowindow" do
     context "ログイン済みの場合" do
       before { sign_in user }
 
       it "InfoWindowを作成する" do
-        post api_infowindow_path, params: {
+        post infowindow_path, params: {
           spot_id: spot.id,
           plan_id: plan.id,
           photo_urls: [ "https://example.com/photo1.jpg" ]
@@ -89,7 +89,7 @@ RSpec.describe "Api::Infowindows", type: :request do
       end
 
       it "place_idで新規スポットを作成してInfoWindowを返す" do
-        post api_infowindow_path, params: {
+        post infowindow_path, params: {
           place_id: "ChIJnew_place",
           lat: 35.6580,
           lng: 139.7016,
@@ -101,7 +101,7 @@ RSpec.describe "Api::Infowindows", type: :request do
       end
 
       it "button_labelを指定できる" do
-        post api_infowindow_path, params: {
+        post infowindow_path, params: {
           spot_id: spot.id,
           button_label: "プランに追加"
         }
@@ -110,7 +110,7 @@ RSpec.describe "Api::Infowindows", type: :request do
       end
 
       it "show_button=falseでボタンを非表示にできる" do
-        post api_infowindow_path, params: {
+        post infowindow_path, params: {
           spot_id: spot.id,
           show_button: "false"
         }
@@ -121,7 +121,7 @@ RSpec.describe "Api::Infowindows", type: :request do
 
     context "未ログインの場合" do
       it "認証エラーを返す" do
-        post api_infowindow_path, params: { spot_id: spot.id }
+        post infowindow_path, params: { spot_id: spot.id }
 
         expect(response).to redirect_to(new_user_session_path)
       end
