+++ spec/system/community_spec.rb
@@ -10,19 +10,19 @@ RSpec.describe "Community", type: :system do
 
   describe "みんなの旅一覧（未ログイン）" do
     it "一覧ページが表示される" do
-      visit plans_path
+      visit community_path
 
       expect(page).to have_content("みんなの旅")
     end
 
     it "検索フォームが表示される" do
-      visit plans_path
+      visit community_path
 
       expect(page).to have_css(".list-page__sidebar")
     end
 
     it "プラン/スポット切り替えタブが表示される" do
-      visit plans_path
+      visit community_path
 
       expect(page).to have_content("プラン")
       expect(page).to have_content("スポット")
